services/image_embedding_service/pyproject.toml
[project]
name = "image-embedding-service"
version = "0.1.0"
description = "Transforms images into embeddings using Hugging Face vision models"
authors = [
    { name = "Tamas Szombathy", email = "szombathy.tamasgergely@gmail.com" }
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "transformers",
    "torch",
    "Pillow",
    "python-multipart"
]
requires-python = ">=3.11.0"

[project.scripts]
image-embedding-service = "image_embedding_service.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["image_embedding_service"]
